src/requirements.txt
aiohttp
http
teams_ai